[chore] separate types file for different resources
