+++ src/types/LaunchDarkly/environment.ts
@@ -1,0 +1,11 @@
+export type Environment = {
+    _id: string
+    key: string
+    name: string
+    color: string
+}
+
+export type Environments = {
+    totalCount: number
+    items: Environment[]
+}
